requirements.txt
fastapi>=0.100.0 # Using a recent version
uvicorn[standard]>=0.20.0 # Includes standard dependencies like watchfiles for reloading 

# For loading .env file
python-dotenv>=1.0.0

# Supabase client library
supabase>=1.0.0 # Using version 1.x 

# Unofficial Garmin Connect library
garminconnect>=0.1.40 # Use a specific recent version 